Refactoring: GeoDataRiverSurvey::MouseShiftRiverPathCenterCommand is moved to separated files.
