+++ libs/geodata/riversurvey/riversurvey.pro
@@ -151,6 +151,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_insertriverpathpointcommand.h \
            private/geodatariversurvey_mouseeditcrosssectioncommand.h \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.h \
+           private/geodatariversurvey_mouseshiftriverpathcentercommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
@@ -203,6 +204,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_insertriverpathpointcommand.cpp \
            private/geodatariversurvey_mouseeditcrosssectioncommand.cpp \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.cpp \
+           private/geodatariversurvey_mouseshiftriverpathcentercommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveycrosssectionwindow_datatabledelegate.cpp \
            private/geodatariversurveycrosssectionwindow_riversurveytabledelegate.cpp \
